Make sure IO::Handle.encoding passes replacement, strict to decoder
The signature had :$replacement and :$strict but they were not passed
through to Encoding.decoder, though they worked fine with the encoder.

In addition add a :$translate-nl argument to allow changing the value.
Previously the default of $translate-nl = True was passed into the
decoder and encoder's but there was no way to configure it.