Give R:I.IntRange its own .push-exactly
This makes HyperIteratorBatcher.produce-batch about 3x as fast as before when
given a Range.iterator..  The dedicated HyperRangeBatcher makes it about 5x as
fast, but introduces a lot of code and maintenance overhead.  Given that the
HyperxBatcher.produce-batch is typically less than 0.1% of total time used,
it feels that just creating an optimized .push-exactly is the right thing to do.