Fix parsing of denormal Nums in &val
We can't just raise 10 into a power to a magnitude of a Num,
as denormals' magnitude would not fit into a normal double.

Fix by using an Int instead and then producing the Num with
nqp::div_In op. This aligns the parsing in &val with the
parsing done in the grammar for literal numbers.