Force a rebuild of compiler and modules on changes to NQP
We use the NQP compiler's new source-digest to detect any changes to NQP.
Changes will trigger a rebuild of rakudo and also affect the Perl 6 compiler's
ID. Thus precomp files will be outdated accordingly and can be recompiled for
the new NQP dependencies.

With this the need for a manual "make clean" after NQP updates and "Missing or
wrong version of dependency" errors in general should be greatly reduced.