Provide child process ID via Proc::Async
The `ready` method now conveys the child process ID. A `pid` method
is added as an alias for `ready`. Arguably, we could just tell people
to use `ready`, but providing a `pid` alias feels like it will be a
little more self-documenting. Sharing a `Promise` between the two as
an implementation detail makes this cheap to do, and we never did
commit to what the `ready` `Promise` evaluates to, so it's safe to
make it end up evaluating to the PID too. So far its use has just
been as a "the process is started" signal, and it will continue to
serve as that just fine.