Make Supply.interval a serial Supply
Previously, it was non-serial, and relied on something downstream
calling `.serialize`. The `whenever` construct would do exactly that,
so the only way to observe this was a `.tap` on the interval `Supply`
directly.

While it's the case that `.tap` itself doesn't (and shouldn't) imply
`.serialize`, it is also the case that built-in `Supply` factories
should give back serial supplies. Somehow, `interval` didn't follow
this rule (likely because that design rule crystalized with time).

Changing this is as likely to fix things as it is to break things (and
has no impact whatsoever on anyone using the `whenever` syntax). As a
bonus, it also makes `whenever Supply.interval($n) { }` perform a
little better, because now the interval `Supply`'s own async lock will
take care of the concurrency control, meaning `.serialize()` is then a
no-op.

Change inspired by a comment on a Stack Overflow question[1] that ran
into the unusual behavior of `Supply.interval`, and reasonably took it
as a behavior that might generalize to all on-demand supplies, when in
fact this case was the outlier.

[1]
https://stackoverflow.com/questions/49539550/can-a-shared-supply-run-multiple-tap-blocks-simultaneously/49544223?noredirect=1#comment86122417_49544223