Use target coercion type in scalar wrap decision
When we compiler a signature with some `T $x`, if T is a type that is
not `Iterable` then we don't need to wrap it in a `Scalar`. This saves
quite some allocations. However, we used the nominal type always to
decide whether to wrap. That's both creating containers when we don't
need to (for example, `Bool()` doesn't need one but would get one),
but also potentially a bug if one coerces to an iterable type from a
non-iterable one (probably odd enough that we didn't hit this issue).