Fix crash in loop + FIRST + LAST when it's wanted
Fixes R#1623 https://github.com/rakudo/rakudo/issues/1623

When we gen Seq.from-loop, we don't anticipate that the body of
the loop was rewritten and there are other bits outside the `while`
op that we need to take care of, so we end up moving the `call` on
var into a separate block, but its declaration remains in original
place, causing the crash observed in the ticket.

Since FIRST-phasered loops already got a block for the loop genned,
annotate the structure with the things we'll need and then when
we end up genning the wanted stage, simply re-use that same block
in the Seq.from-loop, while throwing away no-longer wanted bits,
like the temp var we originally used. Something apparently tries
to access/assign to index 1 of structure in $loop-goods, so we end
up with a `null` at index 0. I couldn't find where that happens,
so I just stuck an empty QAST::Stmts into the first index, to avoid
the `null` creation ('cause `null` causes a crash in MAST gen).

We actually aren't supposed to gen a wanted loop in the particular
case shown in the ticket, since there's no `do` or parens on it. I
filed that separate issue as R#1698
https://github.com/rakudo/rakudo/issues/1698