Support shapes for HAS scoped attributes
The shape declaration on an attributes was parsed correctly since a while,
but now this information is passed to the repr composition. The following
is now supported fo CStructs and CPPStructs:

use NativeCall;
class Foo is repr<CStruct> {
  HAS sometype @.foo[32] is CArray;
}

The memory for the class Foo will be allocate to hold these 32 elements of
type sometype.