Fix crash in double-sink for for ... { Package::foo }
Fixes R#1693 https://github.com/rakudo/rakudo/issues/1693

After adding a sink call to the last statement, we UNWANTED,
the guts and because of the setup, we ended up adding another
sink call into the QAST inside UNWANTED, causing the error in the Issue.

This is really a bandaid and the entire sinkage thing will need
to be re-evaluated and re-worked as a whole, rather than piece-meal.
That's already planned as part of
R#1571 https://github.com/rakudo/rakudo/issues/1571