Create shortcuts for %h.head
It occurred to me that .head on a Map or Hash with only one key, would
be a very fast way to get that one key (and its value).

This commit adds specific .head candidates for Map/Hash and typed Hashes
that are twice as fast as have half of the allocations.  Note that this
is purely an optimization, no new functionality has been added.