Fix wrong error when using module in core namespace
Fixes RT#133107: https://rt.perl.org/Ticket/Display.html?id=133107

My understanding of the cause is that when we try to `require`
stuff we install a package that later gets replacd with the actual thing.
This causes the `is-core` lookup to give a false positive when
the last thing in the chain we look up is that very package.

Fix by checking that HOW of the last thing is not package.