change attributes in role Native from int8 to int
This blew up in module Box2D which is about bindings to a C++ library. The
changed attributes point to not allocated memory. The program crash was
triggered by a method "Box2D::Body.CreateFixture" which heavily involves
traits. So perhaps the error only shows up when a certain level of traits or
mixins are involved.
See https://irclog.perlgeek.de/moarvm/2018-04-11#i_16035620