Use new tryfindmeth op for the sink code-gen
It results in slightly smaller code, saving around 10KB on the compiled
output of CORE.setting. More significantly, it saves a duplicate hash
lookup in the unoptimized path for each sink (previously, we had one
for `can` and one for `findmeth` on the same key). While it was true
that spesh could deal with this in many cases, this gives us better
baseline performance, as well as better performance in megamorphic
cases that spesh cannot help with. In `class C { method m() { } }; for
^100000000 { C.m }`, it saves a fraction of a percent of the cycles,
since that specializes very well. That fraction of a percent is still
3 million cycles overall, however, which will be from the time spent
in the interpreter before the optimized version was ready.