Make `Foo() $x` optimization use nominal type
Since that's all it needs anyway. As noted in issue #1753, subset type
`where` clauses will often not be written so robustly as to handle
having the `Iterable` role passed in to check. Since we've no need to
do that anyway, this change side-steps the problem, in the same way it
is side-stepped for the optimization on a `Foo $x` parameter.