Fix spurious unhandled Failures caused by require
We compiled a require unconditionally to the actual module loading code plus
an indirect lookup of the loaded module. If the loaded comp unit does not
actually contain a module with that name the lookup would fail. However we
don't sink require's return value so the Failure would escape.
Fix by only compiling the indirect lookup when someone's actually interested
in the result.
Seems to fix GH #1332