Make sure to wrap error string in Exception object
The user-facing parts of the Supply API all make sure of this. If we
are directly implementing the Tappable interface, however, then it is
on us to make sure that we only ever call the provided `quit` callback
with an `Exception` object, not a `Str`. This mistake was introduced in
f58ac9991 during an optimization of IO::Socket::Async.

With this, #1775 now reports the error ("connection reset by peer")
correctly, rather than complaining about being unable to throw a
string.