Avoid calling STORE with huge number of arguments for array literals
When initializing array literals like my @a = 1, 2, 3; instead of calling
STORE on the array with a potentially very long argument list, we create code
that pushes those values into a low level array and bind that to the Array's
reifier. This then can lazily set up scalar containers and expand slips.