Fix overzealous X::Syntax::NoSelf throwage
Fixes R#1793 https://github.com/rakudo/rakudo/issues/1793

When we set up a fresh $*HAS_SELF for throwage of X::Syntax::NoSelf in
post contraints, it's not actually scoped to for just the post
constraint. It's technically declared at the start of
variable_declarator and is set only when we get to the place where
we currently set it. This means when we match the shape declaration,
it's entirely unset, which is what is causing the throwage in the Issue.

Fix by moving post_constraint's $*HAS_SELF settage inside the
post_constraint rule.