Re-add $#foo 5to6 error message
Accidentally thrown away in 4ac9915ef66ce054335fccc938a5fc594ec97c24.
We didn't have any test for it in roast master, but there is a test in
v6.c-errata.

The code for `$#` was used for both $# and $#foo feature. The former
is a pre-perl5 variable, and the latter is used frequently in perl5.

I consider this particular 5to6 error message to be useful, so adding
it back (but only in $#foo form).