Make Rationals fully-immutable
Fixes RT#130774: https://rt.perl.org/Ticket/Display.html?id=130774

This commit changes the performance landscape of rationals.

As described in the Work Proposal[^1], the optimization was removed
from +/- of Rationals, so those are now slower, but tossing reduction
calls boosts other areas and lets us use simpler code paths in some ops,
making them faster. So overall the removal of this opt is likely
beneficial (and it had to be removed to fix the data race)

[1] https://github.com/rakudo/rakudo/blob/master/docs/archive/2018-03-04--Polishing-Rationals.md#3-limited-usefulness-of-the-optimization