Look up DEPRECATED at the *end* of compilation
Since traits.pm6 is about the first file in settings being compiled,
it did not know about &DEPRECATED yet.  And somehow it is being forgotten
so the precomp got a VMnull there.  Mitigated by having the coderef for
&DEPRECATED looked up at end of setting compilation by using CHECK.

Finally fixes R #1219 .  skids++ for the nudge to get me looking at this
again.