Move &DEPRECATED to Rakudo::Deprecations.DEPRECATED
The main issue is that in setting compilation, we are referring to
&DEPRECATED *long* before it actually gets defined.  This is normally
not an issue (afaik), but it apparently *is* in the core setting,
especially in combination with precompiled modules doing deprecations.

This provides an alternate fix for 472f6e48444b5ed157658 , and it
paves the way to making "has $.a is DEPRECATED<b>" work.