Make "is DEPRECATED" work on attributes with auto-generated accessors
Tried to do this by adding an ENTER phaser to either generated or already
explictly written accessor methods.  But alas, running $m.add_phaser is
*not* enough if the code for the method has already been compiled: it then
lacks the phaser running code (which it will only generate when it's needed
so not every block checks for phasers at runtime).  Adding the phaser
running code requires access to the QAST, and I don't see that happening
at the moment.