Emit enum to supply on signal
On the recent modification of the signal sub, there was a subtle change
to the form of the signum emitted by the signal handler. Pre-change it
emitted the signum in the form of the corresponding Signal enum and
post-change it emits the signum as an Int. Changing it so that it will
one again emit the Signal enum as it is more useful in varying contexts.
AlexDaniel++ for discovering this.

See [Github Issue 1888](https://github.com/rakudo/rakudo/issues/1888).