Use a MoarVM spesh plugin for private method calls
For those that we can resolve trivially already because the call is
in a class, we simply do so. However, private method calls in roles
have so far been greatly slower, due to a lack of caching and since
the specializer had no idea how to optimize them even if it could.

With this spesh plugin, a module PM.pm6:

    role R {
        method m() { self!p }
        method !p() { 42 }
    }
    class C does R {
    }
    for ^10_000_000 {
        C.m;
    }

When run with `perl6 -I. -e 'use PM6'` will run in 0.83s, compared to
5.5s before, making private method calls in roles around 6.6x faster.

Note that due to the way roles are handled in non-precompiled code,
the rewrite in Perl6::Optimizer (nor any others in there) will not take
effect in a script, thus the module.