Use a spesh plugin for $foo.Qual::Ified::meth()
This lets us optimize away the method resolution after the first time,
which since it has to search in the role structure was quite costly.
The "is the invocant a Qual::Ified" check also gets subsumed by the
guard.

The result is that the program:

    role R1 {
        method m() { 1 }
    }
    role R2 {
        method m() { 2 }
    }
    class C does R1 does R2 {
        method m() {
            self.R2::m()
        }
    }
    for ^10_000_000 {
        C.m
    }

Which used to run in 13.3s now runs in 1.07s, a factor of more than 12x
improvement. This construct appears 66 times in CORE.setting, so there
is a good chance we win something from it there as well as in user code
using it for role disambiguation.