Use a spesh plugin to optimize .?
I explored a couple of different cases. In the first, the program was
written with .?, but in its actual usage the types that showed up were
mostly monomorphic:

    class C {
    }
    class D {
        method m() { 42 }
    }
    for ^10_000_000 {
        (rand > 0.999 ?? C !! D).?m()
    }

This case ran in 10.9s before, and 4.29s after, which is a speedup of
2.5x. Note that the `rand` is a notable cost in this program, so the
speedup to the .? itself is more than that.

Under a polymorphic case:

    class C {
    }
    class D {
        method m() { 42 }
    }
    for ^10_000_000 {
        (rand > 0.5 ?? C !! D).?m()
    }

We go from 7.60s to 4.92s, a 1.5x speedup. Spesh can't just punt this
to doing a deopt for the uncommon case, because there is no uncommon
case. Still, the guard table scan comes out ahead.