Ease maintenance of Actions.nqp, Grammar.nqp
The monolithic files are divided into separate files by class or
grammar. Files are then concatenated into one file, in the required
order, at build time.

Actions.nqp was shortened to the Perl 6 class and the remains divided
into:

+ QActions.nqp
+ RegexGrammar.nqp
+ P5RegexActions.nqp

Grammar.nqp was shortened to the Perl 6 grammar and the remains
divided into:

+ QGrammar.nqp
+ RegexActions.nqp
+ P5RegexGrammar.nqp

The new files were added to the appropriate places in files:

+ tools/build/Makefile-JVM.in
+ tools/build/Makefile-Moar.in