Modify Parameter's positional to rm slurpies
Add slurpy_lol (**@) and slurpy_onearg (+@) to param types that are not
considered positional.

This caused an issue when using '.assuming'. The fact that they were
considered both positional and slurpy caused them
to be added multiple times to a new param list and miscompiled on EVAL
(The below example would result in a redeclaration error).

    sub x(+@a, :$b) { say "{@a.perl} and $b" }; &x.assuming(:!b);

See [GH Issue 1918](https://github.com/rakudo/rakudo/issues/1918).