Add role to Signal for custom zero logic
We want to be able to take advantage of the powerful functionality of
Enumerations, but also implement custom logic in the class. Enter
roles.

Add a role to the Signal class (created via enum keyword earlier in the
unit) to implement custom handling of zero values (to the type object
acting as a sub). jnthn++ for role suggestion.