Make `.?` spesh plugin bug compatible with `.`
The spesh plugin for `.?` fixed a long-standing issue with method
augmented into parent classes not being visible to subclasses. While
that's in theory good, in practice there was code in the wild that
used `.?` in a filtering step and then `.` to call, and inconsistent
behavior between the two (`.?` being fixed and `.` not yet) caused
breakage. This is discussed in issue #1923. Therefore, re-introduce
the bug to `.?`, until we can fix `.` too.