Cope with failure to open a directory in DIR-RECURSE
DIR-RECURSE is used in CompUnit::Repository::FileSystem
S11-repository/curi-install.t installs modules into a directory that
S11-modules/nested.t uses as a CURFS lib. S11-repository/curi-install.t also
deletes its files when it is finished. This can lead to a race condition if
these two tests are run at the same time, but never in isolation.

With this change, we skip any directories that fail to open - this is fine for
how DIR-RECURSE gets used internally.

Many thanks to ugexe++ for uncovering this long standing flapper!
Fixes GH #1956