Helper tool for perf improvement descriptions
This tool should help us consistify commit messages and changelog
entries. It prints the preferred form (e.g. “2x as fast”) to stdout
and some extra variants are printed to stderr along with formulas that
were used. Note that I'm proposing “2x as fast” variant because it is
very easy to calculate and relatively easy to understand. The script
should be modified if we decide to use something else.

The script doesn't handle 0 as new-time, and that's OK because
“∞x as fast” is not very clear anyway.

See also: https://randomascii.wordpress.com/2018/02/04/what-we-talk-about-when-we-talk-about-performance/