Fix WIN32 Proc::Async process termination issue
There is a test on Windows involving Proc::Async and signals that hangs
indefinitely.
See [Github Issue #1975](https://github.com/rakudo/rakudo/issues/1975).

Signals are emulated via libuv on Windows, so the signal handler in the
test is not called. Rather the emulated SIGINT will just call WIN32's
TerminateProcess(). But the process isn't terminated immediately; the
kernel object is not destroyed until other processes with open handles
have released them.

The process is waiting for the handles to be released before it
terminates but Proc::Async is waiting on the process terminating before
it will close it's IO handles (if any open).

Fixing this by calling done on any available suppliers within the kill
method so that the promises are no longer blocked on (in this test
the promises are being awaited). The IO handles are closed and then the
process is terminated.

Note that this solves the issue of the process hanging indefinitely, but
it does not cause the test to pass, since for it to pass the process
would need to receive the signal, execute the signal handler, and send
the right data to stdout.