Naive implementation of uniscript(Str:D)
Inspired by the work of Karl Williamson on Perl 5 as described in:

  https://metacpan.org/pod/distribution/perl/pod/perldelta.pod#Mixed-Unicode-scripts-are-now-detectable

this is a prototype of the basic functionality needed to create this Perl 5
functionality in Perl 6.

This commit creates a function / Cool method "uniscript" that returns the
script of the given String *if* all characters are of the same Unicode Script.
A special case is the "Common" script: if all characters have the "Common"
Script property, then that will be returned.  Otherwise, it will be ignored
in Script comparisons.  If all of the characters have the same same Script
property, then that will be returned.  Otherwise Nil will be returned.
An empty string will be returned for an empty string.

Name / functionality up for bikeshedding.