Str.comb takes its cue from string rather than numeric value
Fixes GH#2010.

In GH#1552 it was decided that the boolean value of allomorphs is decided by
the numerical value, not by the string value.  So from that, it *could* follow
that the behaviour is correct.  However, since `.comb` clearly works on the
the string part of the allomorph, it seemed like the right thing to do to use
the string semantics of Bool instead in this case.