Re-work how we guard Mu-typed assignments
Instead of building up a sizable guard set for the various variables
that may be assigned at that site, we just guard against the Mu type
itself. There may be ways we can be smarter later on, but this should
mean we do rather better with all the code that is generic over many
Scalar containers with different descriptors.