Add missing persons from contributing lists in past announcements
Fixes https://github.com/rakudo/rakudo/issues/2024 R#2024 with the
reasons for missings listed on R#2028
https://github.com/rakudo/rakudo/issues/2028

I've added missing persons with "as well as" clause. When I ran the
contrib script to gen the list for past releases, I had someone
who was on the announcement but not in the script output for pretty
much every release. I suspect these people might've got added/modified
their names in the CREDITS file, but since I'm unable to merge the
original list (sorted by commit count) with the new list, I've added
the missing people list with "as well as" and the missing list is
sorted alphabetically.