Generate less deconts in lowered signature binding
Previously, a method like:

    multi method chars(Str:D: --> Int:D) {
        nqp::p6box_i(nqp::chars($!value))
    }

Would generate 3 `decont` instructions on the invocant. In the event it
was called on a `Str` object, then spesh could eliminate them all. But
if it was called like `$some-scalar-holding-a-str.chars`, there would
be 3 real `decont`s to do. Spesh would lower each one into a fairly
cheap access, but would still have to guard the lookups of each one.
Worse, if it was not a `Scalar`, but a `Proxy`, then it would be 3
`FETCH` calls.

This change gets it down to one `decont` call, which would mean one
instruction to elimiante for a call on `Str`, one to lower and one
guard in the case of `Scalar`, and only one call in the case of
`Proxy`.