Provide p6box_* as a desugar using box_*
Rather than the extops that we have today. For MoarVM we totally get
rid of them. For JVM, the binder being written in Java means we don't
do go quite so far as we still need the code.

This will simplify implementing box/unbox elimination optimization in
MoarVM spesh. Further, it aids our medium-term goal of eliminating the
use of MoarVM extops.