Give PseudoStash its own non-value based .WHICH
PseudoStash is a Map, but a Map that actually can have containers for
values.  Since Map is considered a value-type, this meant that the .WHICH
of e.g. a MY:: could change if the value of a lexical variable was changed.
That is not good.  Instead, mark PseudoStash no longer as a value type, even
though it *is* a Map.