Eradicate Map.STORE_AT_KEY
This was really a Rakudo implementation detail, but was used as a semi-
official API between Map/Hash.  Even though we're now calling .Str on things
that potentially are already Str, the approach of not using a method call
makes Map initialization from a list of values about 1.4x as fast.