Re-instate nqp::null -> Mu return mapping
While we can fix bugs in Rakudo where Perl 6 constructs wrongly produce
it, we have some slangs that use `nqp::ops`, and our hands are rather
tied if we're not wanting to break them. Since we're doing this with a
spesh plugin, the cost is really low anyway (in the best case, it'll be
free), so we might as well put the mapping back and not break a module
that was depending on it.