Avoid building up a huge stack of continuations
If we have a continuation to resume, it already has the continuation
entry frame with error handling as part of the stack that is resumed.
If we thus invoke the continuation inside such a frame, and then it
also results in the continuation being taken, we make the stack more
deeply nested each time. Don't do that. Also refactor to avoid taking
a closure holding the continuation, which will make things a little
cheaper.