make socket-host/-port available on I::S::Async.listen
both are attributes holding a Promise that gets
resolved as soon as the host and port are available,
which is usually immediately.

This allows you to use port 0 for listening. Every new
tap on a .listen will of course get a new port, so
the attributes are on the Tap rather than the Supply.