Add bug compatibility unfix for returning Proxy
Due to a bug in the handling of return values, a Proxy could escape
from a routine that was not marked `is rw`. This was fixed by the
recent spesh plugins work. However, fixing it breaks modules that
depended on the bug, of which there are several. So, for now, bring
the bug back.