Also optimize for ^foo where my constant \foo = 42
Part one of fixing R#2138.  The case for "my constant $foo = 42" either
needs to be fixed differently (as it arrives as a QAST::Var, rather than
a QAST::Want or a QAST::WVal).  Or maybe "my constant $foo = 42" should
be codegenned as a QAST::WVal rather than as a QAST::Var.