Slightly change semantics of (&) on QuantHashes
If the left-hand side of the (&) (also known as ∩) is mutable, then the
resulting QuantHash object will also be mutable.  If it is immutable,
it will stay immutable as before.

This should allow us to better optimize things like:

    my %h is BagHash = a => 42, b => 666;
    my %i is Bag = b => 13, c => 22;
    %h (&)= %i;

This requires quite a few spectest changes.  But all of these tests where
added after Christmas, so we should be good in that respect.

Instigated by https://stackoverflow.com/questions/51580534/does-baggy-add-work-on-mixhash-weights