Fix wrong math with largish Ints
Fixes R#2157 https://github.com/rakudo/rakudo/issues/2157

The cause of the bug is NQP apparently uses doubles to handle its
ints[^1]. 2**52 is the limit under which every number is representable
in a double, which is why when we bumped up[^2] the point at which we
switched to bigints in Rakudo to be over that limit, we brought in the
NQP's limitation into Perl 6 land.

Fix by dialing down the switch to bigint to occur before we go over the
2**52 limit. The NQP Issue[^1] will resolve the problem in NQP land and
once that's done, I presume we can bump up the limit in rakudo once again.

[1] N#363 https://github.com/perl6/nqp/issues/363
[2] https://github.com/rakudo/rakudo/commit/b61b3c7a7f0c6848f4698aaed23da21ca29115bc