Fix for R#2174
This was caused by the erroneous assumption that opcode 10 (vivify role
attribute) of the `BUILDPLAN` would always be the last in the `BUILDPLAN`.
If a `TWEAK` method is available, then the calling of `TWEAK` would follow.
So the execution of `10` opcodes, would run into the run `TWEAK` task,
causing the error.