Tune up Array.ASSIGN-POS
This gets the code size to be quite significantly smaller. It's not
quite at the point where it will inline yet, but a great deal closer.
There's still some suboptimality in the code produced after
specialization, so this will need more attention. Even with these
outstanding tasks, this is still a quite measurable speedup.