Use $*SPEC.dir-sep for separators in CompUnit::RepositoryRegistry
On Windows, keys to the $repos hash are constructed exclusively
with backslash separators, so keys look something like this:

    C:\rakudo\share\perl6\site

However, before this commit, we were using slash characters to construct
paths to look up in that $repos hash, so we were looking for
keys like this as a result:

    C:\rakudo\share\perl6/site

This lookup would fail, and so Windows would end up lacking certain
repositories, such as the "site" repository.

This patch makes sure that $*SPEC.dir-sep is used consistently
throughout RepositoryRegistry to restore these missing repositories to
Windows builds.