Use + prefer full namespace for builder identifier
The reasons for having a hard-coded prefix are outweighed by the need for consistency in how namespaces are declared in meta data. This adds ( and prefers ) a full namespace, and falls back to the previous behavior of prefixing the builder value with `Distribution::Builder::` ( the later behavior can be removed eventually ).