Make sure we call .words on stringified var
By taking out the call to Stringy in the infix:<~> handling, any variable
would be directly exposed to having ".words" called upon it in << >>
post-processing.  This works for IO::Paths, but doesn't work for many
other classes:

  $ perl6 -e 'class A {}; my $a = A.new; dd << foo $a >>'
  No such method 'words' for invocant of type 'A'

Fixed this by inserting a stringification for this particular case of
quotewords processing.  Fixes R#2203