Move some classes out of enclosing methods
We should probably make a pass through and do this for any code that
may be a hot path. I'd hoped we would be able to optimize the closure
captures out easily; alas, we can't, and it can make some otherwise
small and potentially inlineable methods into things that `capturelex`
or similar and so cannot be. For some code that obtained many iterators
of the types involved here, this gave a fairly notable improvement.