Allow prefix calls with .:< ... > on integer literals
Fix Github #2094. Before this commit, '42.:<->' would error because the
check which prevented integers ending with fullstops was overly
aggressive: any time an integer was followed by '.:', an error would be
generated. With this commit, an integer followed by '.:' will only error
if the next thing after it is not '<' then a prefix operator then '>'