Revert "Throws X::Multi::NoMatch if a proto routine is defined without multis."
This reverts commit d0a8a7af0e231a5b31796890c9da390c68433f4d.

The commit breaks an errata spectest. The test can be tweaked or
removed if there's a good justification for doing that, but at the
moment none was presented, so I believe the PR was merged by
mistake. The change can (and likely should) be reintroduced later,
please follow R#2227 for further discussion and progress.

https://github.com/rakudo/rakudo/issues/2227