Don't close over anonymous classes in Seq
Also: move the actual class to Rakudo::Iterator, as we don't want
to pollute the Setting outer lexical namespace further.

Also, now that this lives in Rakudo::Iterator, we should be able to
codegen the call to Rakudo::Iterator.Gather directly, and remove
sub GATHER.  But that's for another time.