Harden deprecation location handling code
It's possible to get an undefined back from `next-interesting-index`,
which caused the slicing to blow up. The undef can happen because, as
in various languages, inlining optimizations result in a backtrace with
with some frames omitted. In the future, we plan to make MoarVM able to
reconstruct a backtrace as if inlining had not taken place; in the
meantime, this avoids the issue reported in #2235.