Restore the "AAS" .. "ABS" behaviour
a949a161446d7698e2 changed the behaviour to be a simple .succ based
iterator (aka, the second element being "AAT").  This restores the
referring to the "..." behaviour, which apparently was TimToady's
original intent.

FWIW, I think this commit should be reverted because the "..." behaviour
it reinstates was even surprising for the current Perl 6 pumpking.

Whatever the outcome, tests will need to be added to codify the behaviour
is the previous commit nor this commit made any spectests break  :-(