Extract Str.words(:autoderef) and streamline it
The runtime post-processing of quoted words used a special candidate
of Str.words.  By giving the method its own name, it is no longer part
of the MMD of Str.words *and* we can remove a named parameter to the
call.  The inside of the method could also be streamlined a bit so that
we have a lot less boxing going on.