Introducing PredictiveIterator role
After some discussion with Jonathan after the SPW, he felt it was better to
move the "count-only" and "bool-only" functionality into a separate role,
which could then be used to smartmatch / MMD on, rather than having to see
whether it is possible to call a method before calling it.

This commit creates that role, with "bool-only" defaulting to booleanizing
the result of "count-only".