At least document the methods that should be supplied
The Positional / Associative roles do not actually enforce the implementation
of needed methods, such as AT-POS / AT-KEY.  Adding stubs for these methods
breaks the setting build, complaining about methods not being implemented
which clearly *are* implemented.

I've tried to add the stubs and defer mixing in the role in a BEGIN block
in the core_epilogue.  Although that compiles ok, the mixed in role is *not*
visible in smart-match (probably nqp::istype), causing all sorts of breakage.
Alas.  Suggestions welcome.