[6.d] Set non-DefiniteHOW default default for DefiniteHOWs
Per 6.d-prep[^1]: On :U/:D type constrained variables, params, attributes,
and parametarized Arrays/Hashes, use the base type as the default
default in 6.d instead of continuing to use the uninitializeable
DefiniteHOW as the type, which prevents use of `.=` and other constructs.

Fixes R#1493. Does not address enums, which is currently blocked by R#2297

[1] https://github.com/perl6/6.d-prep/blob/dffa2642419a0f481591b9b64b602bfc4c4eb66b/TODO/FEATURES.md#make-default-defaults-for-definitehows-be-normal-types
[2] https://github.com/rakudo/rakudo/issues/1493
[3] https://github.com/rakudo/rakudo/issues/2297