Revert "Attempt #2 at making R:I:JSON.from-json faster"
This reverts commit c9432c20721104ced7e2408da5ba1f15ad320975.

Alas, the quick parser accepts JSON that is valid, but does not contain
any curly blocks.  This causes R#2299.  I currently don't see a quick way
to fix this in a way that doesn't leave the way open for other valid JSON
to not be parsed correctly.

I guess we just need to make the NQP regex engine faster.