Improve array assignment inlinability
Prior to this, the ASSIGN-POS method was too large to inline on MoarVM,
but it's really quite helpful for performance if it does inline. Since
MoarVM doesn't factor in the size of nested inlines, however, then it's
possible to get below the inline limit by splitting some of the code out
into a private method, as done by this commit.