Fix wrong type of result in `for 1.5 … 2.5 {}`
The get_bound() coerces to Int, with which we could get away for
opts like `for ^1.5`, but with the new opts, we have to be more
strict in what we accept, or we generate an incorrect sequence.

Fix by typechecking the value we have is an Int, instead of coercing
stuff. This de-optimizes things that where coercion was acceptable
(e.g. doing `for ^1.5`).